Add possibility to configure NIC and CIDR used for tunnel networks
To establish Geneve or Vxlan tunnels there is need to have L3
connectivity with other members of the cluster. This patch adds 2 new
options in the openvswitch API: "tunnelNetworkNic" and
"tunnelNetworkCidr" to allow configuring dedicated network for such
tunnels connectivity.
By default, if those options aren't set, PodIP address will be used as
tunnel endpoint in the OVS POD.