Set nodeName as hostname in ovsdb
Hostname in ovsdb is used by ovn-controller to identify chassis and
should be consistent for all time when POD is running on given node
(ovsdb is stick to the node always) so we can configure
"external-ids:hostname" in ovsdb as nodeName and that way it will always
be the same on the same node and will be easier to identify in e.g.
Neutron API where ovn-controller (neutron agent) really is running.