Add Job Pod to configure OVN controller
There is now additional Pod with Job container created to set some
external-ids in the ovs db for ovn-controller.
Script called init.sh which was called at the beginning of the
ovn-controller container is now moved to that new job container so
reconfiguration of the external ids in ovsdb can be done without
restarting ovn-controller container.