Allow attaching service pods to additional network
For network isolation outgoing traffic, service pods need to be
attached to the network used by the edpm nodes. To achive this, the
`k8s.v1.cni.cncf.io/networks` annotation needs to be set on the
pods. This adds a networkAttachment parameter to the CRD where
it can be specified to which additional network the service pods
should be attached to.

If specified the additional network is used for the OvnEncapIP.

Depends-On: openstack-k8s-operators/lib-common#133