Fix empty OvnEncapIP
On slower systems it seems that the multus network-status annotation
is not set fast enough and won't get passed into the Pod. As a result
the OvnEncapIP is empty:

$ oc logs -p -c ovn-controller ovs-549ql
+ OvnBridge=br-int
+ OvnRemote=tcp:172.17.0.31:6642
+ OvnEncapType=geneve
+ '[' -n internalapi ']'
++ /usr/local/bin/container-scripts/get_net_ip internalapi
Traceback (most recent call last):
  File "/usr/local/bin/container-scripts/get_net_ip", line 63, in <module>
    get_ip_from_network_status()
  File "/usr/local/bin/container-scripts/get_net_ip", line 43, in get_ip_from_network_status
    net_status = json.loads(os.environ.get('PodNetworksStatus'))
  File "/usr/lib64/python3.9/json/__init__.py", line 346, in loads
    return _default_decoder.decode(s)
  File "/usr/lib64/python3.9/json/decoder.py", line 337, in decode
    obj, end = self.raw_decode(s, idx=_w(s, 0).end())
  File "/usr/lib64/python3.9/json/decoder.py", line 355, in raw_decode
    raise JSONDecodeError("Expecting value", s, err.value) from None
json.decoder.JSONDecodeError: Expecting value: line 1 column 1 (char 0)
+ OvnEncapIP=

This changes the daemonset to instead of use the multus network-status
annotation to pass in the nic to be used for the OvnEncapIP and get the
ip information in the init.sh from the nic itself.